tidying up

+++ src/ch_17_vector_and_free_store/ch_17_06_access_to_elements.cpp
@@ -1,53 +1,53 @@
 // program implements an ad-hoc vector class that can be used to create
 // a vector of doubles and demonstrates its usage
 
-#include <iostream> // I/O library header
+#include <iostream> // for cout
 
 // simple class to represent vector of doubles
 // provides a constructor, destructor, get(), set(), and size() methods
 class vector {
     // private members
-    int sz; // size
+    size_t sz; // size
     double *elem; // pointer to the elements
 
-    public:
-        // constructor
-        // create a new vector of s elements
-        explicit vector(int s)
-            : sz {s}, elem {new double[s]} {  }
+public:
+    // constructor
+    // creates a new vector of s elements
+    explicit vector(size_t s)
+        : sz {s}, elem {new double[s]} {  }
 
-        // destructor
-        ~vector()
-        {
-            delete[] elem; // release dynamically allocated resources
-        }
+    // destructor
+    ~vector()
+    {
+        delete[] elem; // release dynamically allocated resources
+    }
 
-        // return the number elements in this vector
-        int size() const
-        {
-            return sz; // current size
-        }
+    // returns the number elements in this vector
+    size_t size() const
+    {
+        return sz; // current size
+    }
 
-        // access: read
-        // return the nth element in this vector
-        double get(int n) const
-        {
-            return elem[n];
-        }
+    // access: read
+    // returns the nth element in this vector
+    double get(size_t n) const
+    {
+        return elem[n];
+    }
 
-        // access: write
-        // set the nth element in this vector to v
-        void set(int n, double v)
-        {
-            elem[n] = v;
-        }
+    // access: write
+    // sets the nth element in this vector to v
+    void set(size_t n, double v)
+    {
+        elem[n] = v;
+    }
 };
 
 int main()
 {
     vector v(5); // build a vector
 
-    for (int i = 0; i < v.size(); i++)
+    for (size_t i = 0; i < v.size(); i++)
     {
         v.set(i, 1.1 + i); // populate elements of v
         // write value of each element in v to stdout
