+++ src/ch_18_vectors_and_arrays/ch_18_05_02_move_semantics.cpp
@@ -1,9 +1,10 @@
 // program implements an ad-hoc class for a vector of doubles and demonstrates
 // how to implement a move constructor and a move assignment operator
 
-#include <iostream> // I/O library header
-#include <algorithm> // algorithm library header
-#include <initializer_list> // provides initializer_list class template
+#include <iostream> // for cout
+#include <algorithm> // for fill(), copy()
+#include <utility> // for move()
+#include <initializer_list> // for initializer_list<>
 
 // ad-hoc class for a vector of doubles
 class vector {
@@ -14,7 +15,7 @@ class vector {
 public:
     // default constructor
     vector()
-        : sz{0}, elem{nullptr} {  }
+        : sz {0}, elem {nullptr} {  }
 
     // public member functions
     // explicit, one-argument constructor
@@ -91,7 +92,7 @@ public:
     // operations for vectors
     // HOWEVER: certain indexing operations will fail for constant vectors since
     // reading from a constant is supported but writing to a constant is illegal
-    double& operator[](int idx) { return elem[idx]; }
+    double& operator[](size_t idx) { return elem[idx]; }
 
     // member access operator for constant vectors
     // provides read access to a vector by returning the value of the
@@ -99,7 +100,7 @@ public:
     // NOTE: function does not provide any range checking
     // MORE IMPORTANTLY: this function is allowed because it returns a value, as
     // opposed to a reference
-    double operator[](int idx) const { return elem[idx]; }
+    double operator[](size_t idx) const { return elem[idx]; }
 
     // destructor deallocates dynamically allocated memory
     ~vector() { delete[] elem; }
