+++ src/ch_20_containers_and_iterators/ch_20_01_01_working_with_data.cpp
@@ -3,15 +3,15 @@
 // dynamically allocated vector of doubles and returns a pointer to it.
 // next the program determines the largest element in each
 
-#include <iostream> // I/O library header
-#include <vector>  // vector library header
+#include <iostream> // for cout
+#include <vector>  // for vector
 
 // jack fills a dynamically allocated array of doubles and returns it. the
 // number of elements contained in the array will be returned through the
 // pointer it receives
 // NOTE: function returns a dynamically allocated pointer to *count doubles, so
 // remember to delete it when done
-double *get_from_jack(int *count);
+double *get_from_jack(size_t *count);
 
 // jill fills a vector and returns a pointer to it
 // NOTE: function returns a dynamically allocated pointer to a vector of
@@ -27,12 +27,12 @@ int main()
     return 0;
 }
 
-double *get_from_jack(int *count)
+double *get_from_jack(size_t *count)
 {
     if (!count)  // return 0 if count is NULL
         return 0;
 
-    const int n = 10;
+    const size_t n = 10;
     double *arr = new double[n];  // allocate array of 10 doubles
 
     if (arr) // if allocation was successful, fill array
@@ -48,7 +48,7 @@ double *get_from_jack(int *count)
 
 std::vector<double> *get_from_jill()
 {
-    const int n = 10;
+    const size_t n = 10;
 
     // allocate pointer to a vector of 10 doubles
     std::vector<double>* arr = new std::vector<double>(n);
@@ -67,7 +67,7 @@ std::vector<double> *get_from_jill()
 
 void fct()
 {
-    int jack_count = 0;
+    size_t jack_count = 0;
 
     // jack will receive a pointer to a dynamically allocated array of doubles,
     // and jack_count will be updated with the number of elements in the
@@ -81,7 +81,7 @@ void fct()
     double *jack_high;
     double *jill_high;
 
-    for (int i = 0; i < jack_count; i++)
+    for (size_t i = 0; i < jack_count; i++)
         if (h < jack_data[i])
         {
             // NOTE: &jack_data[i] == &*(jack_data + i) == (jack_data + i)
