+++ src/ch_20_containers_and_iterators/ch_20_01_02_generalizing_code_02.cpp
@@ -9,15 +9,15 @@
 // program also assigns the pointer to the vector of doubles to a reference to
 // a vector of doubles to simplify accessing it
 
-#include <iostream> // I/O library header
-#include <vector>  // vector library header
+#include <iostream> // for cout
+#include <vector>  // for vector
 
 // jack fills a dynamically allocated array of doubles and returns it. the
 // number of elements contained in the array will be returned through the
 // pointer it receives
 // NOTE: function returns a dynamically allocated pointer to *count doubles, so
 // remember to delete it when done
-double *get_from_jack(int *count);
+double *get_from_jack(size_t *count);
 
 // jill fills a vector and returns a pointer to it
 // NOTE: function returns a dynamically allocated pointer to a vector of
@@ -38,18 +38,18 @@ int main()
     return 0;
 }
 
-double *get_from_jack(int *count)
+double *get_from_jack(size_t *count)
 {
     if (!count) // return 0 if count is NULL
         return 0;
 
-    const int n = 10;
+    const size_t n = 10;
     double *arr = new double[n]; // allocate array of 10 doubles
 
     if (arr) // fill array if allocation was successful
     {
         *count = n; // update count
-        for (int i = 0; i < n; i++) // fill array
+        for (size_t i = 0; i < n; i++) // fill array
             arr[i] = i * 1.1;
     }
 
@@ -58,7 +58,7 @@ double *get_from_jack(int *count)
 
 std::vector<double> *get_from_jill()
 {
-    const int n = 10;
+    const size_t n = 10;
 
     // allocate pointer to vector of 10 doubles
     std::vector<double> *arr = new std::vector<double>(n);
@@ -75,7 +75,7 @@ std::vector<double> *get_from_jill()
 
 double *high(double *first, double *last)
 {
-    double h = -1;
+    double h = -1.0;
     double *high; // pointer for highest element
 
     // iterate over collection
@@ -90,7 +90,7 @@ double *high(double *first, double *last)
 
 void fct()
 {
-    int jack_count = 0;
+    size_t jack_count = 0;
     double *jack_data = get_from_jack(&jack_count);
     std::vector<double> *jill_data = get_from_jill();
